improvements to hoverable panels
hoverable panels now open over the window, so they don't resize and
become jarring to see the transition of.
Speaking of transition, they transition in and out.
Also moved the hoverable panels into a separate file so they could be
maintained easier and referenced in the main file for consistency.